feat: Add persistent mini player with global music playback control
- Create GlobalMusicPlayerCubit for app-wide music state management
- Add GlobalMusicPlayerState to handle playback state across pages
- Implement MiniPlayer widget with play/pause and skip controls
- Add playlist navigation (previous/next song functionality)
- Integrate HydratedBloc for persistent playback state
- Update HomePage to include MiniPlayer at bottom
- Modify song player page to sync with global music state
- Add proper spacing for mini player in all relevant screens
- Implement automatic song progression when track completes
- Add visual feedback with loading states and disabled button states

Key features:
✨ Persistent music playback across app navigation
🎵 Global play/pause control from any screen
⏭️ Skip to next/previous song in playlist
💾 State persistence using HydratedBloc
🎨 Clean UI integration with existing design system
📱 Responsive mini player that appears only when music is loaded

Files modified:
- lib/presentation/music_player/bloc/global_music_player_cubit.dart (new)
- lib/presentation/music_player/bloc/global_music_player_state.dart (new)
- lib/presentation/music_player/widgets/mini_player.dart (new)
- lib/presentation/root/pages/home.dart (updated)
- lib/presentation/song_player/pages/song_player.dart (updated)
- lib/presentation/root/widgets/play_list.dart (updated)
- lib/main.dart (updated)
- lib/service_locator.dart (updated)

+++ lib/presentation/song_player/bloc/song_player_cubit.dart
@@ -80,8 +80,13 @@ class SongPlayerCubit extends HydratedCubit<SongPlayerState> {
   SongPlayerState? fromJson(Map<String, dynamic> json) {
     try {
       currentSongUrl = json['currentSongUrl'];
-      songPosition = Duration(milliseconds: json['positionMillis'] ?? 0);
-      songDuration = Duration(milliseconds: json['durationMillis'] ?? 0);
+      // Fix: Ensure we handle both int and double types
+      songPosition = Duration(
+        milliseconds: (json['positionMillis'] ?? 0).toInt(),
+      );
+      songDuration = Duration(
+        milliseconds: (json['durationMillis'] ?? 0).toInt(),
+      );
 
       return SongPlayerLoaded();
     } catch (_) {
